All packets <received-sent> to the server are now handled properly, we are in-game!

+++ Client/IO/InputQueue.cs
@@ -6,6 +6,6 @@ public sealed class InputQueue : BaseQueue, IConsolidator
     public override ArraySegment<byte> Dequeue(int size) => new(Buffer.Dequeue(size < 0 ? 0 : size));
     public override void Clear() => Buffer.Clear();
     public override int Count => Buffer.Count;
-    public int GetPacketID() => Count >= 1 ? Buffer[0] : -1;
-    public int GetPacketLength() => Count >= 3 ? Buffer[1] << 8 | Buffer[2] : -1;
+    public byte GetPacketID() => (byte)(Count >= 1 ? Buffer[0] : -1);
+    public short GetPacketLength() => (short)(Count >= 3 ? Buffer[1] << 8 | Buffer[2] : -1);
 }
